requirements.txt
streamlit==1.28.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0

